+++ .prettierignore
@@ -2,6 +2,7 @@
 /docs/design/analytics.md
 /dist-schema/
 /goldens/public-api
+/modules/testing/builder/projects/
 /packages/angular_devkit/build_angular/test/
 /packages/angular_devkit/core/src/workspace/json/test/
 /packages/angular_devkit/schematics_cli/blank/project-files/
