+++ modules/testing/builder/projects/hello-world-app/.editorconfig
