+++ modules/testing/builder/projects/hello-world-app/karma.conf.js
