+++ modules/testing/builder/projects/hello-world-app/protractor.conf.js
