+++ modules/testing/builder/projects/hello-world-app/src/favicon.ico
