+++ modules/testing/builder/projects/hello-world-app/src/locale/messages.xlf
