+++ modules/testing/builder/projects/hello-world-app/src/ngsw-config.json
