+++ modules/testing/builder/projects/hello-world-app/src/spectrum.png
