+++ modules/testing/builder/src/index.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+export {
+  BuilderHarness,
+  BuilderHarnessExecutionOptions,
+  BuilderHarnessExecutionResult,
+} from './builder-harness';
+export { HarnessFileMatchers, JasmineBuilderHarness, describeBuilder } from './jasmine-helpers';
+export * from './test-utils';
