+++ packages/angular_devkit/build_angular/BUILD.bazel
@@ -3,10 +3,10 @@
 # Use of this source code is governed by an MIT-style license that can be
 # found in the LICENSE file at https://angular.io/license
 
+load("@npm//@angular/build-tooling/bazel/api-golden:index.bzl", "api_golden_test_npm_package")
 load("@npm//@bazel/jasmine:index.bzl", "jasmine_node_test")
 load("//tools:defaults.bzl", "pkg_npm", "ts_library")
 load("//tools:ts_json_schema.bzl", "ts_json_schema")
-load("@npm//@angular/build-tooling/bazel/api-golden:index.bzl", "api_golden_test_npm_package")
 
 licenses(["notice"])
 
@@ -285,7 +285,6 @@ ts_library(
     testonly = True,
     srcs = glob(
         include = [
-            "src/test-utils.ts",
             "src/testing/**/*.ts",
             "src/**/tests/*.ts",
         ],
@@ -297,6 +296,7 @@ ts_library(
     tsconfig = "//:tsconfig-test.json",
     deps = [
         ":build_angular",
+        "//modules/testing/builder",
         "//packages/angular_devkit/architect",
         "//packages/angular_devkit/architect/node",
         "//packages/angular_devkit/architect/testing",
@@ -394,6 +394,7 @@ LARGE_SPECS = {
             # Dependencies needed to compile and run the specs themselves.
             ":build_angular",
             ":build_angular_test_utils",
+            "//modules/testing/builder",
             "//packages/angular_devkit/architect",
             "//packages/angular_devkit/architect/node",
             "//packages/angular_devkit/architect/testing",
