+++ packages/angular_devkit/build_angular/src/builders/browser/tests/options/named-chunks_spec.ts
@@ -11,7 +11,7 @@ import { BASE_OPTIONS, BROWSER_BUILDER_INFO, describeBuilder } from '../setup';
 
 const MAIN_OUTPUT = 'dist/main.js';
 const NAMED_LAZY_OUTPUT = 'dist/src_lazy-module_ts.js';
-const UNNAMED_LAZY_OUTPUT = 'dist/28.js';
+const UNNAMED_LAZY_OUTPUT = 'dist/358.js';
 
 describeBuilder(buildWebpackBrowser, BROWSER_BUILDER_INFO, (harness) => {
   describe('Option: "namedChunks"', () => {
@@ -61,7 +61,7 @@ describeBuilder(buildWebpackBrowser, BROWSER_BUILDER_INFO, (harness) => {
       const { result } = await harness.executeOnce();
 
       expect(result?.success).toBe(true);
-
+      debugger;
       harness.expectFile(MAIN_OUTPUT).toExist();
       harness.expectFile(NAMED_LAZY_OUTPUT).toNotExist();
       harness.expectFile(UNNAMED_LAZY_OUTPUT).toExist();
