+++ packages/angular_devkit/build_angular/src/testing/index.ts
@@ -6,5 +6,6 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export { BuilderHarnessExecutionOptions, BuilderHarnessExecutionResult } from './builder-harness';
-export { HarnessFileMatchers, describeBuilder } from './jasmine-helpers';
+// TODO: Consider using package.json imports field instead of relative path
+//       after the switch to rules_js.
+export * from '../../../../../modules/testing/builder/src';
